+++ src/paas_charm/expressjs/charm.py
@@ -38,6 +38,7 @@ class Charm(PaasCharm):
 
     Attrs:
         framework_config_class: Base class for framework configuration.
+        paas_config_framework_fields: Mapping from framework fields to paas-config fields.
     """
 
     framework_config_class = ExpressJSConfig
@@ -58,8 +59,8 @@ class Charm(PaasCharm):
         """Return an WorkloadConfig instance."""
         base_dir = pathlib.Path("/app")
         framework_config = typing.cast(ExpressJSConfig, self.get_framework_config())
-        metrics_port, metrics_path, configure_metrics = self._paas_config.metrics_endpoint(
-            default_port=framework_config.port, default_path="/metrics"
+        metrics_port, metrics_path = self._paas_config.metrics_endpoint(
+            default_port=8080, default_path="/metrics"
         )
         return WorkloadConfig(
             framework=self._framework_name,
@@ -72,7 +73,6 @@ class Charm(PaasCharm):
             metrics_target=f"*:{metrics_port}",
             metrics_path=metrics_path,
             metrics_port=metrics_port,
-            configure_metrics=configure_metrics,
             unit_name=self.unit.name,
         )