+++ tests/unit/expressjs/test_app.py
@@ -25,6 +25,8 @@ from paas_charm.redis import PaaSRedisRelationData
             None,
             {
                 "PORT": "8080",
+                "METRICS_PORT": "8080",
+                "METRICS_PATH": "/metrics",
                 "NODE_ENV": "production",
                 "APP_SECRET_KEY": "foobar",
                 "APP_OTHERCONFIG": "othervalue",
@@ -50,6 +52,8 @@ from paas_charm.redis import PaaSRedisRelationData
             ),
             {
                 "PORT": "8080",
+                "METRICS_PORT": "8080",
+                "METRICS_PATH": "/metrics",
                 "NODE_ENV": "production",
                 "APP_SECRET_KEY": "notfoobar",
                 "APP_EXTRA-CONFIG": "extravalue",