+++ tests/unit/flask/test_flask_app.py
@@ -63,7 +63,10 @@ def test_flask_env(
     env = flask_app.gen_environment()
     assert env["FLASK_SECRET_KEY"] == "foobar"
     del env["FLASK_SECRET_KEY"]
-    expected_env = {}
+    expected_env = {
+        "FLASK_METRICS_PORT": "9102",
+        "FLASK_METRICS_PATH": "/metrics",
+    }
     for config_key, config_value in user_defined_config.items():
         if isinstance(config_value, dict):
             for secret_key, secret_value in config_value.items():